+++ source/extensions/filters/network/http_connection_manager/config.h
@@ -61,6 +61,27 @@ public:
 };
 
 /**
+ * Determines if an address is internal based on user provided config.
+ */
+class InternalAddressConfig : public Http::InternalAddressConfig {
+public:
+  InternalAddressConfig(const envoy::config::filter::network::http_connection_manager::v2::
+                            HttpConnectionManager::InternalAddressConfig& config);
+
+  bool isInternalAddress(const Network::Address::Instance& address) const override {
+    if (address.type() == Network::Address::Type::Pipe) {
+      return unix_sockets_;
+    }
+
+    // TODO(snowp): Make internal subnets configurable.
+    return Network::Utility::isInternalAddress(address);
+  }
+
+private:
+  const bool unix_sockets_;
+};
+
+/**
  * Maps proto config to runtime config for an HTTP connection manager network filter.
  */
 class HttpConnectionManagerConfig : Logger::Loggable<Logger::Id::config>,
@@ -94,6 +115,9 @@ public:
   Http::ConnectionManagerStats& stats() override { return stats_; }
   Http::ConnectionManagerTracingStats& tracingStats() override { return tracing_stats_; }
   bool useRemoteAddress() override { return use_remote_address_; }
+  const Http::InternalAddressConfig& internalAddressConfig() const override {
+    return *internal_address_config_;
+  }
   uint32_t xffNumTrustedHops() const override { return xff_num_trusted_hops_; }
   bool skipXffAppend() const override { return skip_xff_append_; }
   const std::string& via() const override { return via_; }
@@ -109,6 +133,7 @@ public:
   Http::ConnectionManagerListenerStats& listenerStats() override { return listener_stats_; }
   bool proxy100Continue() const override { return proxy_100_continue_; }
   const Http::Http1Settings& http1Settings() const override { return http1_settings_; }
+  std::chrono::milliseconds delayedCloseTimeout() const override { return delayed_close_timeout_; }
 
 private:
   typedef std::list<Http::FilterFactoryCb> FilterFactoriesList;
@@ -125,6 +150,7 @@ private:
   Http::ConnectionManagerStats stats_;
   Http::ConnectionManagerTracingStats tracing_stats_;
   const bool use_remote_address_{};
+  const std::unique_ptr<Http::InternalAddressConfig> internal_address_config_;
   const uint32_t xff_num_trusted_hops_;
   const bool skip_xff_append_;
   const std::string via_;
@@ -145,6 +171,7 @@ private:
   Http::DateProvider& date_provider_;
   Http::ConnectionManagerListenerStats listener_stats_;
   const bool proxy_100_continue_;
+  std::chrono::milliseconds delayed_close_timeout_;
 
   // Default idle timeout is 5 minutes if nothing is specified in the HCM config.
   static const uint64_t StreamIdleTimeoutMs = 5 * 60 * 1000;