+++ source/server/configuration_impl.cc
@@ -131,7 +131,9 @@ InitialImpl::InitialImpl(const envoy::config::bootstrap::v2::Bootstrap& bootstra
   admin_.access_log_path_ = admin.access_log_path();
   admin_.profile_path_ =
       admin.profile_path().empty() ? "/var/log/envoy/envoy.prof" : admin.profile_path();
-  admin_.address_ = Network::Address::resolveProtoAddress(admin.address());
+  if (admin.has_address()) {
+    admin_.address_ = Network::Address::resolveProtoAddress(admin.address());
+  }
 
   if (!bootstrap.flags_path().empty()) {
     flags_path_ = bootstrap.flags_path();